Achieve "one-click" pynt-mode
- Jupyter notebook starts up when package is loaded
- pynt-mode creates a new notebook and puts it side-by-side and parses out
  namespaces

This greatly cuts down on the amount of setup work necessary. However we've lost
the custom ability to specify an arbitrary notebook. Honestly we should just do
it automatically.

If there isn't a notebook already side-by-side then create one. But if there
already is one then just use that one. Simple enough.