Make remaining variables buffer-local
For the longest time I made some of the maps global even though intuitively
there was no reason why they should not have been buffer-local.

Now all the core data structures are buffer local and it is simpler to reason
about what variables need to be cleared out when pynt-mode is started on another
buffer (ideally none of them should have to be!).