Fix multiple same file name bug
Notebook switching did not work when visiting two files with the same name
because of the ways emacs changes the name of a buffer after it already exists
to prevent duplicate names. Lesson learned - never cache the name of an emacs
buffer unless you know it's going to be unique.

Also despite my best efforts to keep each code buffer in sync with its notebook,
sometimes it gets inconsistent. I added a keyboard shortcut for you to do that
whenever.