Ostensibly get modified loop unrolling going
I say "ostensibly" because I think there's a bug where a return can get hit
before emacs has had a chance to send all the code cells. Perhaps I can add a
timeout in the top-level exception handler (and I can check the contents of the
exception too to see if it is a return). Or just always wait. I don't think
there's a problem with always waiting before dying...