Make pynt mode more robust
The goal is to get rid of all calls to the function `sit-for' and used deferred
statements instead checking for stuff like window arrangements and variables.
There is only one left! This is making pynt sooo much more robust. I'll actually
feel comfortable recommending it to people to use because of this!