Do not unroll while loops.
It's impossible to do so because there might be a break inside and there's no
way to dump execution to an arbitrary cell in jupyter. We'll handle them similar
to for loops where we won't unroll them but will give the option of unrolling
the first pass to play around with.