Attach to kernel unconditionally
Sure you won't be able to run anything until you actually have a kernel to
attach to. But why would you want to run anything in such an environment anyway?

I guess this makes it pretty bad for when we don't want to attach to a kernel.
Maybe we *should* write a file right before we launch a new notebook to say we
are fine with just a regular environment...