Use heuristic for stopping at our code
Check to see if the next frame's code file starts with a '/'. If so, that
probably means it's code that you pip-installed and it's not local. So don't
keep going down the call stack.

This probably won't work in the general case, but it works for the demo so
that's all the matters for now.